@@ -15,26 +15,24 @@
     },
     "require": {
         "php": ">=7.2.5",
-        "doctrine/doctrine-bundle": "^1.1",
-        "doctrine/orm": "^2.4.8",
+        "doctrine/orm": "^2.6",
+        "doctrine/doctrine-bundle": "2.0",
         "incenteev/composer-parameter-handler": "~2.0",
-        "symfony/dependency-injection": "~3.0",
-        "symfony/swiftmailer-bundle": "~2.6",
-        "symfony/symfony": "3.4",
+        "sensio/framework-extra-bundle":"^5.6.*",
+        "swiftmailer/swiftmailer": "~6.1",
+        "symfony/swiftmailer-bundle": "^3.1",
+        "symfony/symfony": "4.4.*",
         "twig/twig": "^1.0||^2.0"
     },
     "require-dev": {
-        "sensio/generator-bundle": "3.1.7",
-        "symfony/phpunit-bridge": "3.4"
+        "symfony/phpunit-bridge": "~4.0"
     },
     "scripts": {
         "symfony-scripts": [
             "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
+            "bin/console cache:clear --env prod",
+            "bin/console cache:clear",
+            "bin/console assets:install --symlink --relative web"
         ],
         "post-install-cmd": [
             "@symfony-scripts"
@@ -51,11 +49,6 @@
         "sort-packages": true
     },
     "extra": {
-        "symfony-app-dir": "app",
-        "symfony-web-dir": "web",
-        "symfony-var-dir": "var",
-        "symfony-bin-dir": "bin",
-        "symfony-assets-install": "relative",
         "incenteev-parameters": {
             "file": "app/config/parameters.yml"
         },
