@@ -18,9 +18,9 @@ class PagesController extends Controller
 
     public function aboutJsonAction()
     {
-        $json = array('name'  => 'Buty', 'size'  => '32', 'price' => '123.23',);
+        $json = ['name'  => 'Buty', 'size'  => '32', 'price' => '123.23'];
 
-        return new Response(json_encode($json), Response::HTTP_OK, array('Content-type' => 'application/json'));
+        return new Response(json_encode($json), Response::HTTP_OK, ['Content-type' => 'application/json']);
 
     }
 
@@ -31,7 +31,7 @@ class PagesController extends Controller
      */
     public function aboutTwig()
     {
-        return array();
+        return new Response('aboutTwig');
     }
 
     /**
@@ -59,7 +59,7 @@ class PagesController extends Controller
      */
     public function printHeaderAction($title, $color)
     {
-        return array('title' => $title, 'color' => $color,);
+       return $this->render('EduwebTrainingBundle:Pages:printHeader.html.twig', ['title' => $title, 'color' => $color]);
     }
 
     /**
@@ -67,7 +67,7 @@ class PagesController extends Controller
      */
     public function contantPageAction()
     {
-        return $this->forward('EduwebTrainingBundle:Pages:printHeader', array('title' => 'Kontakt', 'color' => 'blue',));
+        return $this->forward('EduwebTrainingBundle:Pages:printHeader', ['title' => 'Kontakt', 'color' => 'blue']);
     }
 
     /**
